refactor zustand store
