+++ ui/src/components/Canvas.tsx
@@ -1016,6 +1016,7 @@ export function Canvas() {
         id,
         parent: "ROOT",
         type: nodetype2dbtype(type),
+        children: [],
         lang: "python",
         x: position.x,
         y: position.y,
@@ -1254,6 +1255,7 @@ export function Canvas() {
         id,
         parent: "ROOT",
         type: "CODE",
+        children: [],
         lang: "python",
         x: position.x,
         y: position.y,
